requirements.txt
streamlit==1.12.0
streamlit_toggle_switch==1.0.2